// src/types/webhook.ts
// Generated by https://quicktype.io

export interface TagEvent {
    event:         string;
    history_items: HistoryItem[];
    task_id:       string;
    webhook_id:    string;
}

export interface HistoryItem {
    id:        string;
    type:      number;
    date:      string;
    field:     string;
    parent_id: string;
    data:      Data;
    source:    null;
    user:      User;
    before:    null;
    after:     After[];
}

export interface After {
    name:    string;
    tag_fg:  string;
    tag_bg:  string;
    creator: number;
}

export interface Data {
}

export interface User {
    id:             number;
    username:       string;
    email:          string;
    color:          string;
    initials:       string;
    profilePicture: null;
}